+++ finagle-toggle/src/test/scala/com/twitter/finagle/toggle/NullToggleMapTest.scala
@@ -1,16 +1,13 @@
 package com.twitter.finagle.toggle
 
-import org.scalacheck.Arbitrary.arbitrary
 import org.scalatest.FunSuite
 import org.scalatestplus.scalacheck.ScalaCheckDrivenPropertyChecks
 
 class NullToggleMapTest extends FunSuite with ScalaCheckDrivenPropertyChecks {
 
-  private val IntGen = arbitrary[Int]
-
   test("apply") {
     val toggle = NullToggleMap("hi")
-    forAll(IntGen) { i => assert(!toggle.isDefinedAt(i)) }
+    assert(toggle.isUndefined)
   }
 
   test("iterator") {