+++ finagle-toggle/src/test/scala/com/twitter/finagle/toggle/ToggleTest.scala
@@ -26,24 +26,25 @@ class ToggleTest extends FunSuite with ScalaCheckDrivenPropertyChecks {
   def newToggle[T](
     id: String,
     pf: PartialFunction[Int, Boolean],
-    fraction: Double
+    fraction: Double,
+    defined: Boolean
   ): Toggle.Fractional = new Toggle.Fractional(id) {
     override def toString: String = s"Toggle($id)"
-    def isDefinedAt(x: Int): Boolean = pf.isDefinedAt(x)
+    def isDefined: Boolean = defined
     def apply(v1: Int): Boolean = pf(v1)
     def currentFraction: Double = fraction
   }
 
   test("True") {
     forAll(IntGen) { i =>
-      assert(on.isDefinedAt(i))
+      assert(on.isDefined)
       assert(on(i))
     }
   }
 
   test("False") {
     forAll(IntGen) { i =>
-      assert(off.isDefinedAt(i))
+      assert(off.isDefined)
       assert(!off(i))
     }
   }
@@ -51,7 +52,7 @@ class ToggleTest extends FunSuite with ScalaCheckDrivenPropertyChecks {
   test("Undefined") {
     val toggle = Toggle.Undefined
     forAll(IntGen) { i =>
-      assert(!toggle.isDefinedAt(i))
+      assert(toggle.isUndefined)
       intercept[UnsupportedOperationException] {
         assert(toggle(i))
       }
@@ -63,7 +64,7 @@ class ToggleTest extends FunSuite with ScalaCheckDrivenPropertyChecks {
     val trueOrTrue = on.orElse(on)
     Seq(trueOrFalse, trueOrTrue).foreach { tog =>
       forAll(IntGen) { i =>
-        assert(tog.isDefinedAt(i))
+        assert(tog.isDefined)
         assert(tog(i))
       }
     }
@@ -72,7 +73,7 @@ class ToggleTest extends FunSuite with ScalaCheckDrivenPropertyChecks {
     val falseOrTrue = off.orElse(on)
     Seq(falseOrFalse, falseOrTrue).foreach { tog =>
       forAll(IntGen) { i =>
-        assert(falseOrFalse.isDefinedAt(i))
+        assert(falseOrFalse.isDefined)
         assert(!falseOrFalse(i))
       }
     }
@@ -82,17 +83,17 @@ class ToggleTest extends FunSuite with ScalaCheckDrivenPropertyChecks {
     val undef1 = Toggle.Undefined
     val undef2 = Toggle.Undefined
     val undef12 = undef1.orElse(undef2)
-    forAll(IntGen) { i => assert(!undef12.isDefinedAt(i)) }
+    assert(undef12.isUndefined)
 
     val t = undef12.orElse(on)
     forAll(IntGen) { i =>
-      assert(t.isDefinedAt(i))
+      assert(t.isDefined)
       assert(t(i))
     }
 
     val f = undef12.orElse(off)
     forAll(IntGen) { i =>
-      assert(f.isDefinedAt(i))
+      assert(f.isDefined)
       assert(!f(i))
     }
   }
@@ -163,21 +164,21 @@ class ToggleTest extends FunSuite with ScalaCheckDrivenPropertyChecks {
   test(
     "Toggle.Fractional.min isDefinedAt is result of isDefinedAt for toggle with lower fraction") {
     val definedAtZero = Toggle.on("com.foo")
-    val undefinedAtZero = newToggle("com.foo", { case x if x > 0 => true }, 1.0)
+    val undefinedAtZero = newToggle("com.foo", { case x if x > 0 => true }, 1.0, false)
 
     val toggle1 = Toggle.Fractional.min(definedAtZero, definedAtZero)
-    assert(toggle1.isDefinedAt(0))
+    assert(toggle1.isDefined)
 
     val toggle2 = Toggle.Fractional.min(undefinedAtZero, definedAtZero)
-    assert(!toggle2.isDefinedAt(0))
+    assert(toggle2.isUndefined)
 
     val toggle3 = Toggle.Fractional.min(definedAtZero, undefinedAtZero)
-    assert(toggle3.isDefinedAt(0))
+    assert(toggle3.isDefined)
   }
 
   test("Toggle.Fractional.min currentFraction is min of both toggles' currentFractions") {
-    val a = newToggle("com.foo", { case _ => true }, 0.5)
-    val b = newToggle("com.foo", { case _ => true }, 0.8)
+    val a = newToggle("com.foo", { case _ => true }, 0.5, defined = true)
+    val b = newToggle("com.foo", { case _ => true }, 0.8, defined = true)
 
     val toggle1 = Toggle.Fractional.min(a, b)
     assert(toggle1.currentFraction == 0.5)
@@ -187,8 +188,8 @@ class ToggleTest extends FunSuite with ScalaCheckDrivenPropertyChecks {
   }
 
   test("Toggle.Fractional.min apply uses currentFraction") {
-    val a = newToggle("com.foo", { case _ => false }, 0.0)
-    val b = newToggle("com.foo", { case _ => true }, 1.0)
+    val a = newToggle("com.foo", { case _ => false }, 0.0, defined = true)
+    val b = newToggle("com.foo", { case _ => true }, 1.0, defined = true)
 
     val toggle1 = Toggle.Fractional.min(a, b)
     assert(toggle1(5) == false)